requirements.txt
python-telegram-bot>=20.0
flask>=3.0
openai>=1.0
google-generativeai>=0.3
httpx>=0.25
python-dotenv>=1.0
pydantic>=2.0

